Update CopybookGeneratorDlg.py
提高兼容性
In the distant past when SWIG was generating wrapper code for C++ static methods it would create a standalone function named ClassName_MethodName for it. When Python added support for static methods then SWIG was able to use that to make a real staticmethod named ClassName.MethodName, but it still generated the standalone function named with the underscore, for compatibility. That underscore version of the static methods is now gone, and you will get an AttributeError in existing code that is using them. To fix the problem simply change the underscore to a dot, for example you should change this:
c = wx.SystemSettings_GetColour(wx.SYS_COLOUR_MENUTEXT)
to this:
c = wx.SystemSettings.GetColour(wx.SYS_COLOUR_MENUTEXT)